requirements.txt
astroid==2.1.0
CacheControl==0.12.5
cached-property==1.5.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
djangorestframework==3.9.1
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
firebase-admin==2.15.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.18.0
idna==2.7
isort==4.3.4
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.6.1
mysqlclient==1.4.2
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.2.2
pytz==2018.9
PyYAML==3.13
requests==2.20.1
rsa==4.0
six==1.12.0
texttable==0.9.1
typed-ast==1.3.1
urllib3==1.24.1
websocket-client==0.54.0
wrapt==1.11.1
